+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [1.532.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.532.0...v1.532.1) (2026-09-03)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#2936](https://github.com/adobe/spacecat-audit-worker/issues/2936)) ([eadae22](https://github.com/adobe/spacecat-audit-worker/commit/eadae22c1f5bddd39e5f3b9d4be21e76801f1336))
+
+# [1.532.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.531.0...v1.532.0) (2026-09-02)
+
+
+### Features
+
+* **brand-claims:** forward enrichment mode on the ready-signal (LLMO-7312) ([#2937](https://github.com/adobe/spacecat-audit-worker/issues/2937)) ([75b982c](https://github.com/adobe/spacecat-audit-worker/commit/75b982c4c8fac8121611dbc94615f65620c2aa9c)), closes [experience-platform/mystique#4804](https://github.com/experience-platform/mystique/issues/4804)
+
+# [1.531.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.530.1...v1.531.0) (2026-09-02)
+
+
+### Features
+
+* **brand-claims:** on-demand one-shot trial runs (LLMO-7263) ([#2932](https://github.com/adobe/spacecat-audit-worker/issues/2932)) ([aebc00a](https://github.com/adobe/spacecat-audit-worker/commit/aebc00a561c50fe2c3ce96c488976bb1fc36943d))
+
+## [1.530.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.530.0...v1.530.1) (2026-09-01)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#2931](https://github.com/adobe/spacecat-audit-worker/issues/2931)) ([3a34cc8](https://github.com/adobe/spacecat-audit-worker/commit/3a34cc8eec0815af66f528f0f84db506ddd8f297))
+
 # [1.530.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.529.5...v1.530.0) (2026-08-31)